fix: register bank precompile
