+++ precompile/contracts/bank/bank_test.go
@@ -1,57 +1,55 @@
 package bank_test
 
-// import (
-// 	"encoding/hex"
-// 	"math/big"
-// 	"testing"
+import (
+	"encoding/hex"
+	"math/big"
+	"testing"
 
-// 	"github.com/CosmWasm/wasmd/app"
-// 	"github.com/CosmWasm/wasmd/precompile/contracts/bank"
-// 	"github.com/CosmWasm/wasmd/precompile/registry"
-// 	"github.com/cosmos/cosmos-sdk/crypto/hd"
-// 	"github.com/cosmos/evm/x/vm/statedb"
-// 	"github.com/cosmos/go-bip39"
-// 	"github.com/ethereum/go-ethereum/common"
-// 	"github.com/ethereum/go-ethereum/core/vm"
-// 	"github.com/ethereum/go-ethereum/crypto"
-// 	"github.com/stretchr/testify/require"
+	"github.com/CosmWasm/wasmd/app"
+	"github.com/CosmWasm/wasmd/precompile/contracts/bank"
+	"github.com/cosmos/cosmos-sdk/crypto/hd"
+	"github.com/cosmos/evm/x/vm/core/vm"
+	"github.com/cosmos/evm/x/vm/statedb"
+	"github.com/cosmos/go-bip39"
+	"github.com/ethereum/go-ethereum/common"
+	"github.com/ethereum/go-ethereum/crypto"
+	"github.com/stretchr/testify/require"
 
-// 	sdkmath "cosmossdk.io/math"
-// 	cryptotypes "github.com/cosmos/cosmos-sdk/crypto/types"
-// 	sdk "github.com/cosmos/cosmos-sdk/types"
-// 	authztypes "github.com/cosmos/cosmos-sdk/x/authz"
-// 	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
-// 	evmtypes "github.com/cosmos/evm/x/vm/types"
-// )
+	sdkmath "cosmossdk.io/math"
+	cryptotypes "github.com/cosmos/cosmos-sdk/crypto/types"
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
+	evmtypes "github.com/cosmos/evm/x/vm/types"
+)
 
-// func MockAddressPair() (sdk.AccAddress, common.Address) {
-// 	return PrivateKeyToAddresses(MockPrivateKey())
-// }
+func MockAddressPair() (sdk.AccAddress, common.Address) {
+	return PrivateKeyToAddresses(MockPrivateKey())
+}
 
-// func MockPrivateKey() cryptotypes.PrivKey {
-// 	// Generate a new Sei private key
-// 	entropySeed, _ := bip39.NewEntropy(256)
-// 	mnemonic, _ := bip39.NewMnemonic(entropySeed)
-// 	algo := hd.Secp256k1
-// 	derivedPriv, _ := algo.Derive()(mnemonic, "", "")
-// 	return algo.Generate()(derivedPriv)
-// }
+func MockPrivateKey() cryptotypes.PrivKey {
+	// Generate a new Sei private key
+	entropySeed, _ := bip39.NewEntropy(256)
+	mnemonic, _ := bip39.NewMnemonic(entropySeed)
+	algo := hd.Secp256k1
+	derivedPriv, _ := algo.Derive()(mnemonic, "", "")
+	return algo.Generate()(derivedPriv)
+}
 
-// func PrivateKeyToAddresses(privKey cryptotypes.PrivKey) (sdk.AccAddress, common.Address) {
-// 	// Encode the private key to hex (i.e. what wallets do behind the scene when users reveal private keys)
-// 	testPrivHex := hex.EncodeToString(privKey.Bytes())
+func PrivateKeyToAddresses(privKey cryptotypes.PrivKey) (sdk.AccAddress, common.Address) {
+	// Encode the private key to hex (i.e. what wallets do behind the scene when users reveal private keys)
+	testPrivHex := hex.EncodeToString(privKey.Bytes())
 
-// 	// Sign an Ethereum transaction with the hex private key
-// 	key, _ := crypto.HexToECDSA(testPrivHex)
-// 	msg := crypto.Keccak256([]byte("foo"))
-// 	sig, _ := crypto.Sign(msg, key)
+	// Sign an Ethereum transaction with the hex private key
+	key, _ := crypto.HexToECDSA(testPrivHex)
+	msg := crypto.Keccak256([]byte("foo"))
+	sig, _ := crypto.Sign(msg, key)
 
-// 	// Recover the public keys from the Ethereum signature
-// 	recoveredPub, _ := crypto.Ecrecover(msg, sig)
-// 	pubKey, _ := crypto.UnmarshalPubkey(recoveredPub)
+	// Recover the public keys from the Ethereum signature
+	recoveredPub, _ := crypto.Ecrecover(msg, sig)
+	pubKey, _ := crypto.UnmarshalPubkey(recoveredPub)
 
-// 	return sdk.AccAddress(privKey.PubKey().Address()), crypto.PubkeyToAddress(*pubKey)
-// }
+	return sdk.AccAddress(privKey.PubKey().Address()), crypto.PubkeyToAddress(*pubKey)
+}
 
 // func TestSend(t *testing.T) {
 // 	denom := "orai"
@@ -231,44 +229,55 @@ package bank_test
 // 	require.Equal(t, grantCoin.Amount, grantCoins[0].Amount.Sub(burnCoins[0].Amount))
 // }
 
-// func TestBalance(t *testing.T) {
-// 	denom := "orai"
-// 	tApp := app.Setup(t)
-// 	ctx := tApp.NewContext(true)
-// 	mockAddr, mockEVMAddr := MockAddressPair()
-// 	tApp.EvmKeeper.SetAddressMapping(ctx, mockAddr, mockEVMAddr)
-// 	sdk.RegisterDenom(denom, sdkmath.LegacyNewDec(6))
-// 	mintCoins := sdk.NewCoins(sdk.NewCoin(denom, sdkmath.NewInt(100000)))
-// 	sentCoins := sdk.NewCoins(sdk.NewCoin(denom, sdkmath.NewInt(10)))
-// 	bankKeeper := tApp.GetBankKeeper()
-// 	accountKeeper := tApp.GetAccountKeeper()
-// 	authzKeeper := tApp.GetAuthzKeeper()
-// 	err := bankKeeper.MintCoins(ctx, evmtypes.ModuleName, mintCoins)
-// 	require.NoError(t, err)
-// 	tApp.GetBankKeeper().SendCoinsFromModuleToAccount(ctx, evmtypes.ModuleName, mockAddr, sentCoins)
-// 	tApp.GetBankKeeper().SetParams(ctx, banktypes.DefaultParams())
+func TestBalance(t *testing.T) {
+	denom := "orai"
+	tApp := app.Setup(t)
+	ctx := tApp.NewContext(true)
+	mockAddr, mockEVMAddr := MockAddressPair()
+	tApp.EvmKeeper.SetAddressMapping(ctx, mockAddr, mockEVMAddr)
+	sdk.RegisterDenom(denom, sdkmath.LegacyNewDec(6))
+	mintCoins := sdk.NewCoins(sdk.NewCoin(denom, sdkmath.NewInt(100000)))
+	sentCoins := sdk.NewCoins(sdk.NewCoin(denom, sdkmath.NewInt(10)))
+	bankKeeper := tApp.GetBankKeeper()
+	err := bankKeeper.MintCoins(ctx, evmtypes.ModuleName, mintCoins)
+	require.NoError(t, err)
+	tApp.GetBankKeeper().SendCoinsFromModuleToAccount(ctx, evmtypes.ModuleName, mockAddr, sentCoins)
+	tApp.GetBankKeeper().SetParams(ctx, banktypes.DefaultParams())
 
-// 	evm := vm.EVM{
-// 		StateDB: statedb.New(ctx, tApp.EvmKeeper, statedb.NewEmptyTxConfig(common.BytesToHash(ctx.HeaderHash()))),
-// 	}
-// 	p := bank.NewContract(tApp.EvmKeeper, bankKeeper, accountKeeper, authzKeeper)
-// 	method := bank.ABI.Methods[bank.BalanceMethod]
-// 	suppliedGas := uint64(10_000_000)
+	// Set EVM parameters to register the bank precompile address
+	EVMParams := tApp.GetEVMKeeper().GetParams(ctx)
+	EVMParams.ActiveStaticPrecompiles = append(EVMParams.ActiveStaticPrecompiles, bank.BankPrecompileAddress)
+	tApp.GetEVMKeeper().SetParams(ctx, EVMParams)
+	p, found, err := tApp.GetEVMKeeper().GetPrecompileInstance(ctx, common.HexToAddress(bank.BankPrecompileAddress))
 
-// 	args, err := method.Inputs.Pack(mockEVMAddr, denom)
-// 	require.Nil(t, err)
-// 	res, _, err := p.Run(&evm, registry.AddrContractAddress, registry.AddrContractAddress,
-// 		append(method.ID, args...),
-// 		suppliedGas,
-// 		false,
-// 		nil,
-// 	)
-// 	require.Nil(t, err)
-// 	output, err := method.Outputs.Unpack(res)
-// 	require.Nil(t, err)
-// 	require.Equal(t, 1, len(output))
-// 	require.Equal(t, output[0].(*big.Int), big.NewInt(sentCoins[0].Amount.Int64()))
-// }
+	require.True(t, found)
+	require.NoError(t, err)
+
+	contract := p.Map[common.HexToAddress(bank.BankPrecompileAddress)]
+	require.NotNil(t, contract)
+
+	evm := vm.EVM{
+		StateDB: statedb.New(ctx, tApp.EvmKeeper, statedb.NewEmptyTxConfig(common.BytesToHash(ctx.HeaderHash()))),
+	}
+	method := bank.ABI.Methods[bank.BalanceMethod]
+	suppliedGas := uint64(10_000_000)
+
+	args, err := method.Inputs.Pack(mockEVMAddr, denom)
+	require.Nil(t, err)
+	res, _, err := evm.RunPrecompiledContract(
+		contract,
+		vm.AccountRef(mockEVMAddr),
+		append(method.ID, args...),
+		suppliedGas,
+		nil,
+		false,
+	)
+	require.Nil(t, err)
+	output, err := method.Outputs.Unpack(res)
+	require.Nil(t, err)
+	require.Equal(t, 1, len(output))
+	require.Equal(t, output[0].(*big.Int), big.NewInt(sentCoins[0].Amount.Int64()))
+}
 
 // func TestSupply(t *testing.T) {
 // 	denom := "orai"
